T: Refactor RsExtractTraitTest
